+++ www/src/components/App.js
@@ -1,13 +1,14 @@
 import React from 'react'
-import Header from './Header'
-import Router from './Router'
+import Header from './header/Header.js'
+// import Page from './page/Page'
+// import Footer from './footer/Footer'
 import './App.css'
 
 const App = () => (
   <div>
     <Header />
-    <Router />
+    {/* <Page />
+    <Footer /> */}
   </div>
 )
 export default App
-
