--- www/src/components/Header.css
@@ -1,7 +1,0 @@
-
-.nav-link {
-    cursor: pointer!important;
-}
-.navbar-brand {
-    cursor: pointer!important;
-}
