+++ www/src/components/old/Header.css
@@ -1,0 +1,7 @@
+
+.nav-link {
+    cursor: pointer!important;
+}
+.navbar-brand {
+    cursor: pointer!important;
+}
