--- www/src/rxdux/RxComponent.js
@@ -1,15 +1,0 @@
-import React, {Component, PropTypes} from 'react'
-
-class RxComponent extends Component {
-  render () {
-    return (
-      <div />
-    )
-  }
-}
-
-RxComponent.propTypes = {
-
-}
-
-export default RxComponent
